third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-activeElement-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
This tests the activeElement property of a ShadowRoot.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
